+++ NEWS.md
@@ -1,11 +1,13 @@
 # AirMonitor 0.3.12
 
-* Updated `monitor_loadLatest()` and `monitor_loadDaily()` with improved logic to handle multiple
+* Updated `monitor_loadLatest()`, `monitor_loadDaily()` and `monitor_loadAnnual()`
+with improved logic to handle multiple
 deployments at a single location. These functions intentionally only return a 
 single `deviceDeploymentID` per `locationID` and preferentially retain AirNow
 data over data from AIRSIS or WRCC. The new logic applies to cases where multiple
 deployments exist within AirNow (or AIRSIS or WRCC) and retains the 
 deployment that has the most recent data.
+* Addressed CRAN package documentation issue.
 
 # AirMonitor 0.3.11